Add files via upload
The LED Flasher project, which makes use of an Arduino UNO  board with 12 LEDs and different patterns, has illustrated the adaptability and  creative possibilities of systems based on microcontrollers. We have successfully  built a dynamic lighting display that can create compelling visual displays by  utilizing the strength of Arduino and the simplicity of LED components.